split off HcalCaloFlagTool, it's rarely used
