+++ DataFormats/METReco/interface/HcalCaloFlagLabels.h
@@ -1,12 +1,12 @@
-#ifndef GUARD_HCALCALOFLAGLABELS_H
-#define GUARD_HCALCALOFLAGLABELS_H
+#ifndef DataFormats_METReco_HcalCaloFlagLabels_h
+#define DataFormats_METReco_HcalCaloFlagLabels_h
 
 #include "DataFormats/HcalDetId/interface/HcalSubdetector.h"
 #include <string>
 
 // Create alias names for all status bits
 // These aliases are valid for only the _current release_
-// Use the HcalCaloFlagTool (below) for full interpretation
+// Use the HcalCaloFlagTool for full interpretation
 namespace HcalCaloFlagLabels
 {
   //subdetector-specific bits defined here (bits 0-15)
@@ -59,22 +59,4 @@ namespace HcalCaloFlagLabels
   
 }
 
-/** \brief Provides interpretation of flag bits with understanding of 
-    CMSSW version dependence.
-*/
-class HcalCaloFlagTool {
-public:
-  HcalCaloFlagTool(const std::string& releaseName);
-  std::string getFieldName(HcalSubdetector sd, int bit) const;
-  int getFieldWidth(HcalSubdetector sd, int bit) const;
-  int getFieldStart(HcalSubdetector sd, const std::string& field) const;
-  int getFieldWidth(HcalSubdetector sd, const std::string& field) const;
-  bool hasField(HcalSubdetector sd, const std::string& field) const;
-  bool hasField(HcalSubdetector sd, int bit) const;
-private:
-  std::string releaseName_;
-  bool standardFormat_;
-  int major_, minor_, patch_, subpatch_;
-};
-
-#endif
+#endif //DataFormats_METReco_HcalCaloFlagLabels_h
