+++ DataFormats/METReco/src/HcalCaloFlagTool.cc
@@ -1,4 +1,4 @@
-#include "DataFormats/METReco/interface/HcalCaloFlagLabels.h"
+#include "DataFormats/METReco/interface/HcalCaloFlagTool.h"
 #include <string.h>
 #include <cstdio>
